Изменил структуру
