+++ src/components/app/app.jsx
@@ -1,11 +1,11 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import {Switch, Route, BrowserRouter} from 'react-router-dom';
-import MainScreen from '../main-screen/main-screen';
-import FavoritesScreen from '../favorites-screen/favorites-screen';
-import LoginScreen from '../login-screen/login-screen';
-import OfferScreen from '../offer-screen/offer-screen';
-import NotFoundScreen from '../not-found-screen/not-found-screen';
+import MainScreen from '../pages/main-screen/main-screen';
+import FavoritesScreen from '../pages/favorites-screen/favorites-screen';
+import LoginScreen from '../pages/login-screen/login-screen';
+import OfferScreen from '../pages/offer-screen/offer-screen';
+import NotFoundScreen from '../pages/not-found-screen/not-found-screen';
 
 const App = ({offerCount}) => {
   return (
