+++ src/components/pages/favorites-screen/favorites-screen.jsx
@@ -1,32 +1,11 @@
 import React from 'react';
-import {Link} from 'react-router-dom';
+import Footer from '../../wrappers/footer/footer';
+import Header from '../../wrappers/header/header';
 
 const FavoritesScreen = () => {
   return (
     <React.Fragment>
-      <header className="header">
-        <div className="container">
-          <div className="header__wrapper">
-            <div className="header__left">
-              <Link className="header__logo-link" to="/">
-                <img className="header__logo" src="img/logo.svg" alt="6 cities logo" width="81" height="41" />
-              </Link>
-            </div>
-            <nav className="header__nav">
-              <ul className="header__nav-list">
-                <li className="header__nav-item user">
-                  <a className="header__nav-link header__nav-link--profile" href="#">
-                    <div className="header__avatar-wrapper user__avatar-wrapper">
-                    </div>
-                    <span className="header__user-name user__name">Oliver.conner@gmail.com</span>
-                  </a>
-                </li>
-              </ul>
-            </nav>
-          </div>
-        </div>
-      </header>
-
+      {<Header />}
       <main className="page__main page__main--favorites">
         <div className="page__favorites-container container">
           <section className="favorites">
@@ -153,12 +132,7 @@ const FavoritesScreen = () => {
           </section>
         </div>
       </main>
-      <footer className="footer container">
-        <Link className="footer__logo-link" to="/">
-          <img className="footer__logo" src="img/logo.svg" alt="6 cities logo" width="64" height="33" />
-        </Link>
-      </footer>
-
+      {<Footer />}
     </React.Fragment>
   );
 };
