+++ src/components/pages/login-screen/login-screen.jsx
@@ -1,31 +1,10 @@
 import React from 'react';
-import {Link} from 'react-router-dom';
+import Header from '../../wrappers/header/header';
 
 const LoginScreen = () => {
   return (
     <React.Fragment>
-      <header className="header">
-        <div className="container">
-          <div className="header__wrapper">
-            <div className="header__left">
-              <Link className="header__logo-link" to="/">
-                <img className="header__logo" src="img/logo.svg" alt="6 cities logo" width="81" height="41" />
-              </Link>
-            </div>
-            <nav className="header__nav">
-              <ul className="header__nav-list">
-                <li className="header__nav-item user">
-                  <a className="header__nav-link header__nav-link--profile" href="#">
-                    <div className="header__avatar-wrapper user__avatar-wrapper">
-                    </div>
-                    <span className="header__login">Sign in</span>
-                  </a>
-                </li>
-              </ul>
-            </nav>
-          </div>
-        </div>
-      </header>
+      {<Header />}
       <main className="page__main page__main--login">
         <div className="page__login-container container">
           <section className="login">
