+++ src/components/pages/main-screen/main-screen.jsx
@@ -1,6 +1,7 @@
 import React from 'react';
-import PlaceCard from '../place-card';
+import PlaceCard from '../../place-card';
 import PropTypes from 'prop-types';
+import Header from '../../wrappers/header/header';
 
 
 const cardsCount = [1, 2, 3, 4, 5];
@@ -9,29 +10,7 @@ const MainScreen = ({offerCount}) => {
 
   return (
     <React.Fragment>
-      <header className="header">
-        <div className="container">
-          <div className="header__wrapper">
-            <div className="header__left">
-              <a className="header__logo-link header__logo-link--active">
-                <img className="header__logo" src="img/logo.svg" alt="6 cities logo" width="81" height="41" />
-              </a>
-            </div>
-            <nav className="header__nav">
-              <ul className="header__nav-list">
-                <li className="header__nav-item user">
-                  <a className="header__nav-link header__nav-link--profile" href="#">
-                    <div className="header__avatar-wrapper user__avatar-wrapper">
-                    </div>
-                    <span className="header__user-name user__name">Oliver.conner@gmail.com</span>
-                  </a>
-                </li>
-              </ul>
-            </nav>
-          </div>
-        </div>
-      </header>
-
+      {<Header />}
       <main className="page__main page__main--index">
         <h1 className="visually-hidden">Cities</h1>
         <div className="tabs">
